v1.0.3 Modification of control interface for Robotiq grippers
Based on the package from [waypointrobotics/robotiq_85_gripper](https://github.com/waypointrobotics/robotiq_85_gripper), a control interface of both Robotiq grippers (140mm and 85mm) was developed, allowing to control the grippers through the use of an Action Client.
Whats new: 
- Action Client-Server structure developed for control of multiple robotiq grippers.
- Publication of real-time gripper joint state.

This packages should migrate in the future to an independent repository or to the robotiq ros-industrial repository.